requirements.txt
alembic==1.16.1
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
cssselect2==0.8.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.3.1
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.0
SQLAlchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
typing_extensions==4.13.2
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
